+++ core/tests/taskhook_test.rs
